585: Moved OrderTypeSelector to MakeWidget subcomponents
