+++ src/styled-components/WidgetHeader/WidgetHeader.tsx
@@ -6,7 +6,7 @@ export const WidgetHeader = styled.div`
   display: flex;
   justify-content: space-between;
   min-height: 2rem;
-  margin-bottom: 2rem;
+  margin-bottom: 1.875rem;
   width: 100%;
 
   @media ${breakPoints.phoneOnly} {
